requirements.txt
aiosqlite==0.18.0
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
databases==0.7.0
exceptiongroup==1.1.1
fastapi==0.95.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.2.2
python-dotenv==1.0.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
